// src/components/common/Tours/tours/getStarted/enums.ts
export enum GetStartedTourTargets {
  WelcomeButton = 'getStartedWelcome',
  EnterTitleField = 'getStartedEnterTitle',
  ChooseActionButton = 'getStartedChooseAction',
  SelectActionButton = 'getStartedSelectAction',
  CustomizeColonyFields = 'getStartedCustomizeColony',
  SelectLogoField = 'getStartedSelectLogo',
  EnterDescriptionField = 'getStartedEnterDescription',
  SelectDecisionMethod = 'getStartedSelectDecisionMethod',
  ConfirmActionButton = 'getStartedConfirmAction',
}